+++ src/core/ports/users_cache.go
@@ -5,6 +5,6 @@ import integration "go-api/src/infra/http/integrations/jsonplaceholder/responses
 type (
 	UsersCache interface {
 		Set(key string, value []integration.User, timeEx int)
-		Get(key string) []integration.User
+		Get(key string) ([]integration.User, error)
 	}
 )