+++ src/main/app.go
@@ -2,6 +2,7 @@ package app
 
 import (
 	"go-api/src/main/container"
+	"go-api/src/main/module/amqp/rabbitmq"
 	database "go-api/src/main/module/db/mysql"
 	http_server "go-api/src/main/module/http/server"
 	"go-api/src/main/module/work"
@@ -12,19 +13,21 @@ import (
 
 type Server struct {
 	Container *container.Container
-	engine    http_server.HttpServer
+	http      http_server.HttpServer
 	db        database.Database
+	amqp      rabbitmq.RabbitMq
 }
 
 func (server *Server) Setup() *Server {
 	server.Container = container.NewContainer(
 		container.NewContainerConfig(server.db.Db),
 	)
-	work := work.New(server.Container)
 
-	work.StartCrons()
+	work.New(server.Container).StartCrons()
 
-	server.engine.New(server.Container)
+	go server.amqp.New(server.Container).Start()
+
+	server.http.New(server.Container)
 
 	return server
 }