+++ src/main/module/amqp/amqp-server.go
@@ -1,0 +1,15 @@
+package amqp_server
+
+import (
+	"go-api/src/main/container"
+	consumer "go-api/src/presentation/amqp/consumers"
+)
+
+type AmqpServer interface {
+	New(c *container.Container) AmqpServer
+	Start()
+	StartConsumers(container []consumer.Comsumer, i int)
+	NeedToReconnect(err error, msg string)
+	LoadConsumers(c *container.Container) []consumer.Comsumer
+	LoadProducers(c *container.Container)
+}