+++ src/main/module/cache/redis/client.go
@@ -3,6 +3,8 @@ package redis
 import (
 	"context"
 	cache_client "go-api/src/main/module/cache"
+	"log"
+	"os"
 	"time"
 
 	"github.com/go-redis/redis/v8"
@@ -10,22 +12,29 @@ import (
 
 type redisClient struct {
 	redis *redis.Client
+	ctx   context.Context
 }
 
-var ctx = context.Background()
-
 func New() cache_client.CacheClient {
+	var ctx = context.Background()
+
 	rdb := redis.NewClient(&redis.Options{
-		Addr: "localhost:6379",
+		Addr: os.Getenv("REDIS_HOST") + ":" + os.Getenv("REDIS_PORT"),
 	})
 
-	return &redisClient{redis: rdb}
+	log.Default().Print("Redis: Connection with sucessfully")
+	return &redisClient{redis: rdb, ctx: ctx}
 }
 
 func (rc *redisClient) Get(key string) (string, error) {
-	return rc.redis.Get(ctx, key).Result()
+	result, err := rc.redis.Get(rc.ctx, key).Result()
+	if err == redis.Nil {
+		return "", nil
+	}
+
+	return result, nil
 }
 
 func (rc *redisClient) Set(key string, value string, timeEx int) error {
-	return rc.redis.Set(ctx, key, value, time.Duration(time.Duration(timeEx).Milliseconds())).Err()
+	return rc.redis.Set(rc.ctx, key, value, time.Duration(time.Duration(timeEx).Milliseconds())).Err()
 }