+++ src/presentation/amqp/ports/message.go
@@ -1,0 +1,5 @@
+package ports_amqp
+
+type Message struct {
+	Body interface{}
+}